Errors display improvements
