+++ webapp/app/session/errors/route.js
@@ -1,10 +1,22 @@
-import PaginatedFilteredRoute from '../../routes/paginated_filtered_route';
+import BaseRoute from '../../routes/base';
+import AuthenticatedRouteMixin from 'ember-simple-auth/mixins/authenticated-route-mixin';
 
-export default PaginatedFilteredRoute.extend({
+export default BaseRoute.extend(AuthenticatedRouteMixin, {
 
-    needs: ['session'],
 
-    model: function(params) {
-        return this.store.query('error', {session_id: this.modelFor('session').id, page: params.page || 1});
+    model: function() {
+        let session = this.modelFor('session');
+
+        return this.store.query('error', {session_id: session.id});
+    },
+
+    setupController: function(controller, model) {
+        this._super(controller, model);
+        controller.set('session', this.modelFor('session'));
+        controller.set('errors', model);
+    },
+
+    renderTemplate: function() {
+        this.render('errors');
     }
 });
