+++ webapp/app/session/errors/template.hbs
@@ -1,5 +1,1 @@
-{{#each model as |error| }}
-  {{error-details err=error}}
-{{/each}}
-
-{{pagination-nav page=page pages_total=pages_total next_page="next_page" prev_page="prev_page" goto_page="goto_page"}}
+{{outlet}}
