+++ webapp/app/session/template.hbs
@@ -1,95 +1,1 @@
-{{#tool-bar}}
-  <form class="navbar-form navbar-right">
-    {{# if (or session.data.authenticated.user_info.can.moderate session.data.authenticated.user_info.can.admin) }}
-      <button class="btn" {{action "toggle_archive"}}><i class="fa fa-archive"></i> {{if session_model.archived 'Unarchive' 'Archive'}}</button>
-
-      {{# if needs_investigation }}
-        <button class="btn btn-default" {{action "start_investigating"}}><i class="fa fa-check"></i> Mark as investigated</button>
-
-      {{else}}
-          {{# if session_model.investigated }}
-            <button class="btn btn-danger" {{action "toggle_investigated"}}><i class="fa fa-times"></i> Mark not Investigated</button>
-          {{/if}}
-      {{/if}}
-    {{/if}}
-
-
-  </form>
-  <p class="navbar-text">Session {{session_model.display_id}}</p>
-
-{{/tool-bar}}
-
-
-{{#if investigating}}
-  {{#modal-dialog close="cancel_investigating"
-     targetAttachment="center"
-     translucentOverlay=true}}
-    <p>You are marking this session as <strong>investigted</strong></p>
-    <div class="container-fluid">
-      {{textarea value=investigate_conclusion cols="80" rows="6" class="form-control" placeholder="Investigation conclusion (e.g. 'Test bug')"}}
-    </div>
-    <div class="container-fluid add-margin-top">
-      {{#if investigate_conclusion}}
-      <button class="btn btn-default pull-right add-margin-left" {{action "finish_investigating"}}>Submit</button>
-      {{/if}}
-      <button class="btn btn-danger pull-right add-margin-left" {{action "cancel_investigating"}}>Cancel</button>
-    </div>
-  {{/modal-dialog}}
-{{/if}}
-
-
-{{hr-separator title="Overview"}}
-
-<div class="container-fluid session-details">
-  <div class="col-md-2">
-    {{#link-to "user.sessions" session_model.user_email}}
-      <img class="img-circle" title="Started by {{session.user_email}}" src="{{gravatar-url session_model.user_email}}" />
-    {{/link-to}}
-  </div>
-  <div class="col-md-8">
-    {{session-breakdown session=session_model}}
-    <table class="table details">
-      <tr>
-        <td class="faint">Run time: </td>
-        <td>
-          <strong>{{moment-time unix=session_model.start_time}}</strong> &rarr;
-          <strong>{{moment-time unix=session_model.end_time}}</strong>
-        </td>
-      </tr>
-      <tr>
-        <td class="faint">Test Subjects:</td>
-        <td>
-          {{#each session_model.subjects as |subject|}}
-            <i class="fa fa-barcode"></i>
-            <strong>{{subject.name}}</strong> ({{subject.product}} v{{subject.version}}, r{{subject.revision}})
-          {{/each}}
-        </td>
-      </tr>
-    </table>
-
-  </div>
-</div>
-
-{{hr-separator title="Tests"}}
-
-<div class="container-fluid">
-  {{query-results results=tests filter_action=(action "filter") goto_page=(action "goto_tests_page") goto_test=(action "goto_test")}}
-</div>
-
-{{hr-separator title="Additional Information"}}
-
-<div class="container-fluid">
-{{#if metadata.slash.commandline}}
-  Command-line
-  <div class="console">
-    {{metadata.slash.commandline}}
-  </div>
-{{/if}}
-</div>
-
-
-{{hr-separator title="Activity"}}
-
-{{activity/activity-timeline items=history}}
-
-{{activity/new-comment}}
+{{outlet}}
