+++ webapp/app/styles/components/errors/traceback-frame.scss
@@ -38,10 +38,12 @@
 
     margin: 0.5em;
     border: 1px solid $gray-light;
+    color: $gray-light;
 }
 
 .vars-heading {
     background: $backslash-orange-lighter;
+    
     font-size: small;
 }
