--- webapp/app/templates/errors-failures-notice.hbs
@@ -1,9 +1,0 @@
-{{#if obj.num_errors}}
-  <div class="row">
-    <div class="col-xs-12">
-      <div class="alert alert-danger">
-        {{title-case obj.typename}} has errors!
-      </div>
-    </div>
-  </div>
-{{/if}}
