+++ webapp/app/test/errors/controller.js
@@ -3,5 +3,6 @@ import Ember from 'ember';
 
 export default Ember.Controller.extend({
 
-    test: Ember.computed.alias('parent_controller.test')
+    test: Ember.computed.oneWay('parent_controller.test'),
+    session: Ember.computed.oneWay('test.session')
 });
