+++ webapp/app/test/single_error/route.js
@@ -1,6 +1,7 @@
 import Ember from 'ember';
+import AuthenticatedRouteMixin from 'ember-simple-auth/mixins/authenticated-route-mixin';
 
-export default Ember.Route.extend({
+export default Ember.Route.extend(AuthenticatedRouteMixin, {
 
     model: function(params) {
 
@@ -12,11 +13,12 @@ export default Ember.Route.extend({
         this._super(controller, model);
         controller.set('error', model);
         controller.set('test', this.controllerFor('test').get('test'));
-        
+        controller.set('session', this.controllerFor('test').get('test.session'));
+
     },
 
     renderTemplate: function() {
-        this.render('error');
+        this.render('single_error');
     }
 
 });
